requirements.txt
django-cms>=3.6,<3.7
djangocms-admin-style>=1.3,<1.5
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=3.7,<3.8
djangocms-link>=2.3
djangocms-style>=2.1
djangocms-googlemap>=1.2
djangocms-snippet>=2.1,<2.2
djangocms-video>=2.0,<2.2
djangocms-column>=1.9
djangocms-file>=2.2,<3.0
djangocms-picture>=2.0,<2.2
easy_thumbnails
django-filer>=1.3
Django<2.0
django-classy-tags>=0.9
django-sekizai>=1.0
django-mptt>0.9
html5lib>=1.0.1
Pillow>=3.0
six
pytz
mysqlclient
django-extensions